requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.6
django-cors-headers==3.13.0
django-filter==22.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
Markdown==3.4.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
